chore: distribute navigation code
